Finished abstract GenerativeController and created skeletons for the main UniformSimulation. Updated Aiport and Runway interface to provide more useful methods. Fixed totals and max delays in sim and temp stats from double -> int.

+++ lib/simulation/concretes/simulation_stats.dart
@@ -1,15 +1,13 @@
-// simulation_stats.dart
-
 import 'package:air_traffic_sim/simulation/concretes/temp_stats.dart';
 
-/// stores the metrics of a simulation
+/// Stores the metrics of a simulation
 class SimulationStats {
   final double averageLandingDelay;
   final double averageHoldTime;
   final double averageDepartureDelay;
   final double averageWaitTime;
-  final double maxLandingDelay;
-  final double maxDepartureDelay;
+  final int maxLandingDelay;
+  final int maxDepartureDelay;
   final int maxInboundQueue;
   final int maxOutboundQueue;
   final int totalCancellations;
@@ -52,8 +50,8 @@ class SimulationStats {
       averageHoldTime: 0.0,
       averageDepartureDelay: 0.0, 
       averageWaitTime: 0.0,
-      maxLandingDelay: 0.0, 
-      maxDepartureDelay: 0.0,
+      maxLandingDelay: 0, 
+      maxDepartureDelay: 0,
       maxInboundQueue: 0,
       maxOutboundQueue: 0,
       totalCancellations: 0,
